use ament_package_xml to handle deps move read_all and relax to drqp_control as bins
