+++ packages/drqp_control/package.xml
@@ -12,6 +12,7 @@
   <depend>ros2_controllers</depend>
   <depend>rclcpp</depend>
   <depend>std_msgs</depend>
+  <depend>drqp_a1_16_driver</depend>
   <!-- <depend>rqt_joint_trajectory_controller</depend> --> <!-- https://github.com/ros-controls/ros2_controllers/tree/master/rqt_joint_trajectory_controller -->
