+++ packages/drqp_control/src/ReadAll.cpp
@@ -18,7 +18,10 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-#include "DrQp.h"
+#include <iostream>
+
+#include "drqp_control/DrQp.h"
+
 #include "drqp_a1_16_driver/XYZrobotServo.h"
 #include "drqp_serial/TcpSerial.h"
 #include "drqp_serial/UnixSerial.h"
@@ -28,9 +31,23 @@ int main()
   // TcpSerial servoSerial("192.168.1.136", 2022);
   UnixSerial servoSerial("/dev/ttySC0");
 
-  for (const auto servoId : servoIdsRange()) {
-    XYZrobotServo servo(servoSerial, servoId);
-    servo.torqueOff();
+  for (const auto& leg : kAllLegServoIds) {
+    int servoIndexInLeg = 0;
+    for (const int servoId : leg) {
+      XYZrobotServo servo(servoSerial, servoId);
+
+      XYZrobotServoStatus status = servo.readStatus();
+      if (servo.isFailed()) {
+        std::cerr << legNameForServo(servoId) << " servo: " << servoId
+                  << " error reading status: " << servo.getLastError() << "\n";
+      } else {
+        std::cout << legNameForServo(servoId) << "\tservo: " << servoId << ": " << status.position
+                  << "\t neutral: " << kNeutralPose[kServoIdToLeg[servoId]][servoIndexInLeg]
+                  << "\n";
+      }
+      ++servoIndexInLeg;
+    }
+    std::cout << "\n";
   }
 
   return 0;
